Enum pallet_nbv_storage::pallet::Call

pub enum Call<T: Config> {

    set_xpub {
        xpub: BoundedVec<u8, T::XPubLen>,
    },
    remove_xpub {},
    create_vault {
        threshold: u32,
        description: BoundedVec<u8, T::VaultDescriptionMaxLen>,
        include_owner_as_cosigner: bool,
        cosigners: BoundedVec<T::AccountId, T::MaxCosignersPerVault>,
    },
    remove_vault {
        vault_id: [u8; 32],
    },
    propose {
        vault_id: [u8; 32],
        recipient_address: BoundedVec<u8, T::XPubLen>,
        amount_in_sats: u64,
        description: BoundedVec<u8, T::VaultDescriptionMaxLen>,
    },
    remove_proposal {
        proposal_id: [u8; 32],
    },
    set_bdk_url {
        new_url: BoundedVec<u8, ConstU32<32>>,
    },
    save_psbt {
        proposal_id: [u8; 32],
        signature_payload: BoundedVec<u8, T::PSBTMaxLen>,
    },
    finalize_psbt {
        proposal_id: [u8; 32],
        broadcast: bool,
    },
    broadcast_psbt {
        proposal_id: [u8; 32],
    },
    kill_storage {},
    ocw_insert_descriptors {
        payload: VaultsPayload<T::Public>,
        signature: T::Signature,
    },
    ocw_insert_psbts {
        payload: ProposalsPayload<T::Public>,
        signature: T::Signature,
    },
    ocw_finalize_psbts {
        payload: ProposalsPayload<T::Public>,
        signature: T::Signature,
    },
    // some variants omitted

}

Contains one variant per dispatchable that can be called by an extrinsic.

Variants

set_xpub

Fields

xpub: BoundedVec<u8, T::XPubLen>

XPub insertion

This extrinsic inserts a user-defined xpub as well as in the pallet storage.

Parameters:

xpub: Extended public key, it can be sent with or without fingerprint/derivation path

Considerations

The origin must be Signed and the sender must have sufficient funds free for the transaction fee.

This extrinsic is marked as transactional, so if an error is fired, all the changes will be reverted (but the fees will be applied nonetheless).

This extrinsic cannot handle a xpub update (yet). if it needs to be updated, remove it first and insert a new one.

remove_xpub

Fields

Xpub removal

Removes the linked xpub from the account which signs the transaction. The xpub will be removed from both the pallet storage and identity registration.

This tx does not takes any parameters.

create_vault

Fields

threshold: u32

description: BoundedVec<u8, T::VaultDescriptionMaxLen>

include_owner_as_cosigner: bool

cosigners: BoundedVec<T::AccountId, T::MaxCosignersPerVault>

Vault insertion

Inserts the vault on chain. Meant to be used by an offchain worker.

Parameters:

threshold: The number of signatures needed for a proposal to be approved/finalized

description: A small definition. What will the vault be used for?

include_owner_as_cosigner: Add automatically the owner as cosigner

cosigners: The other accounts that will participate in vault proposals.

Considerations

Do not include the vault owner on the cosigners list.

remove_vault

Fields

vault_id: [u8; 32]

Vault removal

Tries to remove vault and all its proposals, only the owner can call this extrinsic.

Parameters:

vault_id: the vault to be removed with all its proposals

Considerations:

Only the vault owner can perform this extrinsic

propose

Fields

vault_id: [u8; 32]

recipient_address: BoundedVec<u8, T::XPubLen>

amount_in_sats: u64

description: BoundedVec<u8, T::VaultDescriptionMaxLen>

Vault transaction proposal

Inserts a proposal on the specified vault.

Parameters:

vault_id: the vault identifier in which the proposal will be inserted

recipient_address: Mainnet address to which the funds will be send

amount_in_sats: Amount to send in satoshis.

description: The reason for the proposal, why do you are proposing this?.

Considerations

Please ensure the recipient address is a valid mainnet address.

remove_proposal

Fields

proposal_id: [u8; 32]

Proposal removal

Tries to remove a specified proposal. Only the user who created the proposal can remove it.

Parameters:

proposal_id: the proposal identifier

set_bdk_url

Fields

new_url: BoundedVec<u8, ConstU32<32>>

BDK URL insertion

Changes the BDK-services endpoint, useful for pointing to the btc mainnet or testnet

Parameters:

new_url: The new endpoint to which all the bdk related requests will be sent.

Considerations

Ensure the new url is valid.

The url has a maximum length of 32 bytes

save_psbt

Fields

proposal_id: [u8; 32]

signature_payload: BoundedVec<u8, T::PSBTMaxLen>

PSBT signature insertion

Stores the signature for a PSBT proposal

Parameters:

proposal_id: the proposal identifier

signature_payload: a blob of psbt bytes, resulting from a external wallet

Considerations

If successful, this process cannot be undone

A user can only sign a proposal once

finalize_psbt

Fields

proposal_id: [u8; 32]

broadcast: bool

Finalize PSBT

Queries a proposal to be finalized generating a tx_id in the process, it can also be broadcasted if specified.

Parameters:

proposal_id: the proposal identifier

broadcast: A boolean flag

Considerations

If successful, this process cannot be undone

The proposal must have a valid PSBT

Any vault member can perform this extrinsic

broadcast_psbt

Fields

proposal_id: [u8; 32]

Broadcast PSBT

Queries a proposal to be broadcasted in case it wasn’t on the finalization step.

Parameters:

proposal_id: the vault identifier in which the proposal will be inserted

Considerations

If successful, this process cannot be undone

The proposal must be finalized already

Any vault member can perform this extrinsic

kill_storage

Fields

Kill almost all storage

Use with caution!

Can only be called by root and removes All vaults and proposals

ocw_insert_descriptors

Fields

payload: VaultsPayload<T::Public>

signature: T::Signature

Extrinsic to insert a valid vault descriptor

Meant to be unsigned with signed payload and used by an offchain worker

ocw_insert_psbts

Fields

payload: ProposalsPayload<T::Public>

signature: T::Signature

Extrinsic to insert a valid proposal PSBT

Meant to be unsigned with signed payload and used by an offchain worker

ocw_finalize_psbts

Fields

payload: ProposalsPayload<T::Public>

signature: T::Signature

Extrinsic to insert a valid proposal TX_ID

Meant to be unsigned with signed payload and used by an offchain worker

Implementations

impl<T: Config> Call<T>

pub fn new_call_variant_set_xpub(xpub: BoundedVec<u8, T::XPubLen>) -> Self

Create a call with the variant set_xpub.

pub fn new_call_variant_remove_xpub() -> Self

Create a call with the variant remove_xpub.

pub fn new_call_variant_create_vault(
threshold: u32,
description: BoundedVec<u8, T::VaultDescriptionMaxLen>,
include_owner_as_cosigner: bool,
cosigners: BoundedVec<T::AccountId, T::MaxCosignersPerVault>
) -> Self

Create a call with the variant create_vault.

pub fn new_call_variant_remove_vault(vault_id: [u8; 32]) -> Self

Create a call with the variant remove_vault.

pub fn new_call_variant_propose(
vault_id: [u8; 32],
recipient_address: BoundedVec<u8, T::XPubLen>,
amount_in_sats: u64,
description: BoundedVec<u8, T::VaultDescriptionMaxLen>
) -> Self

Create a call with the variant propose.

pub fn new_call_variant_remove_proposal(proposal_id: [u8; 32]) -> Self

Create a call with the variant remove_proposal.

pub fn new_call_variant_set_bdk_url(
new_url: BoundedVec<u8, ConstU32<32>>
) -> Self

Create a call with the variant set_bdk_url.

pub fn new_call_variant_save_psbt(
proposal_id: [u8; 32],
signature_payload: BoundedVec<u8, T::PSBTMaxLen>
) -> Self

Create a call with the variant save_psbt.

pub fn new_call_variant_finalize_psbt(
proposal_id: [u8; 32],
broadcast: bool
) -> Self

Create a call with the variant finalize_psbt.

pub fn new_call_variant_broadcast_psbt(proposal_id: [u8; 32]) -> Self

Create a call with the variant broadcast_psbt.

pub fn new_call_variant_kill_storage() -> Self

Create a call with the variant kill_storage.

pub fn new_call_variant_ocw_insert_descriptors(
payload: VaultsPayload<T::Public>,
signature: T::Signature
) -> Self

Create a call with the variant ocw_insert_descriptors.

pub fn new_call_variant_ocw_insert_psbts(
payload: ProposalsPayload<T::Public>,
signature: T::Signature
) -> Self

Create a call with the variant ocw_insert_psbts.

pub fn new_call_variant_ocw_finalize_psbts(
payload: ProposalsPayload<T::Public>,
signature: T::Signature
) -> Self

Create a call with the variant ocw_finalize_psbts.